requirements.txt
httplib2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-debug-toolbar==1.4
django-storages==1.1.8
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
djangorestframework==3.3.1
gunicorn==19.3.0
Markdown==2.6.5
Pillow==3.0.0
psycopg2==2.6.1
sqlparse==0.1.16
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.4
